rpc: move GetCoinInfo and ListDynamicFields to alpha service
