+++ crates/sui-e2e-tests/tests/rpc/coin_info.rs
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use sui_macros::sim_test;
-use sui_rpc_api::proto::node::v2::node_service_client::NodeServiceClient;
-use sui_rpc_api::proto::node::v2::GetCoinInfoRequest;
-use sui_rpc_api::proto::node::v2::GetCoinInfoResponse;
+use sui_rpc_api::proto::node::v2alpha::node_service_client::NodeServiceClient;
+use sui_rpc_api::proto::node::v2alpha::GetCoinInfoRequest;
+use sui_rpc_api::proto::node::v2alpha::GetCoinInfoResponse;
 use sui_sdk_types::TypeTag;
 use test_cluster::TestClusterBuilder;
