+++ crates/sui-rpc-api/proto/sui.node.v2alpha.proto
@@ -8,6 +8,7 @@ syntax = "proto3";
 package sui.node.v2alpha;
 
 import "sui.node.v2.proto";
+import "sui.types.proto";
 
 // Service for subscribing to data from a Sui Fullnode
 service SubscriptionService {
@@ -41,4 +42,122 @@ message SubscribeCheckpointsResponse {
 
   // The requested data for this checkpoint
   optional sui.node.v2.GetFullCheckpointResponse checkpoint = 2;
+}
+
+// Service for reading data from a Sui Full node.
+service NodeService {
+  // Request information for the specified coin type.
+  rpc GetCoinInfo(GetCoinInfoRequest) returns (GetCoinInfoResponse);
+
+  // List the dynamic fields for provided parent.
+  rpc ListDynamicFields(ListDynamicFieldsRequest) returns (ListDynamicFieldsResponse);
+}
+
+// Request message for `NodeService.GetCoinInfo`.
+message GetCoinInfoRequest {
+  // The coin type to request information about
+  optional sui.types.TypeTag coin_type = 1;
+}
+
+// Response message for `NodeService.GetCoinInfo`.
+message GetCoinInfoResponse {
+  // Required. The coin type.
+  optional sui.types.TypeTag coin_type = 1;
+
+  // Optional. This field will be populated with information about this coin
+  // type's `0x2::coin::CoinMetadata` if it exists and has not been wrapped.
+  optional CoinMetadata metadata = 2;
+
+  // Optional. This field will be populated with information about this coin
+  // type's `0x2::coin::TreasuryCap` if it exists and has not been wrapped.
+  optional CoinTreasury treasury = 3;
+
+  // TODO add support for tracking RegulatedCoinMetadata
+  // Optional. If this coin type is a regulated coin, this field will be
+  // populated with information about its `0x2::coin::RegulatedCoinMetadata`
+  // object.
+  // optional RegulatedCoinMetadata regulated_metadata = 4;
+}
+
+// Metadata for a coin type
+message CoinMetadata {
+  // ObjectId of the `0x2::coin::CoinMetadata` object.
+  optional sui.types.ObjectId id = 1;
+  // Number of decimal places to coin uses.
+  optional uint32 decimals = 2;
+  // Name for the token
+  optional string name = 3;
+  // Symbol for the token
+  optional string symbol = 4;
+  // Description of the token
+  optional string description = 5;
+  // URL for the token logo
+  optional string icon_url = 6;
+}
+
+// Information about a coin type's `0x2::coin::TreasuryCap` and its total available supply
+message CoinTreasury {
+  // ObjectId of the `0x2::coin::TreasuryCap` object.
+  optional sui.types.ObjectId id = 1;
+  // Total available supply for this coin type.
+  optional uint64 total_supply = 2;
+}
+
+// Information about a regulated coin, which indicates that it makes use of the transfer deny list.
+message RegulatedCoinMetadata {
+  // ObjectId of the `0x2::coin::RegulatedCoinMetadata` object.
+  optional sui.types.ObjectId id = 1;
+  // The ID of the coin's `CoinMetadata` object.
+  optional sui.types.ObjectId coin_metadata_object = 2;
+  // The ID of the coin's `DenyCap` object.
+  optional sui.types.ObjectId deny_cap_object = 3;
+}
+
+// Request message for `NodeService.ListDynamicFields`
+message ListDynamicFieldsRequest {
+  // Required. The `UID` of the parent, which owns the collections of dynamic fields.
+  optional sui.types.ObjectId parent = 1;
+
+  // The maximum number of dynamic fields to return. The service may return fewer than this value.
+  // If unspecified, at most `50` entries will be returned.
+  // The maximum value is `1000`; values above `1000` will be coerced to `1000`.
+  optional uint32 page_size = 2;
+
+  // A page token, received from a previous `ListDynamicFields` call.
+  // Provide this to retrieve the subsequent page.
+  //
+  // When paginating, all other parameters provided to `ListDynamicFields` must
+  // match the call that provided the page token.
+  optional string page_token = 3;
+}
+
+// Response message for `NodeService.ListDynamicFields`
+message ListDynamicFieldsResponse {
+  // Page of dynamic fields owned by the specified parent.
+  repeated DynamicField dynamic_fields = 1;
+
+  // A token, which can be sent as `page_token` to retrieve the next page.
+  // If this field is omitted, there are no subsequent pages.
+  optional string next_page_token = 2;
+}
+
+message DynamicField {
+  // Required. ObjectId of this dynamic field's parent.
+  optional sui.types.ObjectId parent = 1;
+
+  // Required. ObjectId of this dynamic field.
+  optional sui.types.ObjectId field_id = 2;
+
+  // Required. The type of the dynamic field "name"
+  optional sui.types.TypeTag name_type = 3;
+
+  // Required. The serialized move value of "name"
+  optional bytes name_value = 4;
+
+  // Optional. The ObjectId of the child object when a child is a dynamic
+  // object field.
+  //
+  // The presence or absence of this field can be used to determine if a child
+  // is a dynamic field or a dynamic child object
+  optional sui.types.ObjectId dynamic_object_id = 5;
 }
