+++ crates/sui-rpc-api/src/grpc.rs
@@ -279,27 +279,6 @@ impl crate::proto::node::v2::node_service_server::NodeService for crate::RpcServ
             .map(tonic::Response::new)
             .map_err(Into::into)
     }
-
-    async fn get_coin_info(
-        &self,
-        request: tonic::Request<crate::proto::node::v2::GetCoinInfoRequest>,
-    ) -> Result<tonic::Response<crate::proto::node::v2::GetCoinInfoResponse>, tonic::Status> {
-        self.get_coin_info(request.into_inner())
-            .map(tonic::Response::new)
-            .map_err(Into::into)
-    }
-
-    async fn list_dynamic_fields(
-        &self,
-        request: tonic::Request<crate::proto::node::v2::ListDynamicFieldsRequest>,
-    ) -> std::result::Result<
-        tonic::Response<crate::proto::node::v2::ListDynamicFieldsResponse>,
-        tonic::Status,
-    > {
-        self.list_dynamic_fields(request.into_inner())
-            .map(tonic::Response::new)
-            .map_err(Into::into)
-    }
 }
 
 use crate::proto::node::v2alpha::SubscribeCheckpointsResponse;
@@ -415,4 +394,28 @@ fn apply_checkpont_options(
     }
 
     checkpoint
+}
+
+#[tonic::async_trait]
+impl crate::proto::node::v2alpha::node_service_server::NodeService for crate::RpcService {
+    async fn get_coin_info(
+        &self,
+        request: tonic::Request<crate::proto::node::v2alpha::GetCoinInfoRequest>,
+    ) -> Result<tonic::Response<crate::proto::node::v2alpha::GetCoinInfoResponse>, tonic::Status> {
+        self.get_coin_info(request.into_inner())
+            .map(tonic::Response::new)
+            .map_err(Into::into)
+    }
+
+    async fn list_dynamic_fields(
+        &self,
+        request: tonic::Request<crate::proto::node::v2alpha::ListDynamicFieldsRequest>,
+    ) -> std::result::Result<
+        tonic::Response<crate::proto::node::v2alpha::ListDynamicFieldsResponse>,
+        tonic::Status,
+    > {
+        self.list_dynamic_fields(request.into_inner())
+            .map(tonic::Response::new)
+            .map_err(Into::into)
+    }
 }
